Agregamos Mi lista como nueva funcionlidad y las practicas de JS

+++ routes/web.php
@@ -4,14 +4,23 @@ Route::get('/', 'MovieController@home');
 
 Route::get('/movies', 'MovieController@index');
 Route::get('/movies/new', 'MovieController@new')->middleware(['auth', 'is_admin']);
-Route::post('/movies/new', 'MovieController@create');
-Route::delete('/movies/{id}', 'MovieController@delete');
+Route::post('/movies/new', 'MovieController@create')->middleware(['auth', 'is_admin']);
+Route::delete('/movies/{id}', 'MovieController@delete')->middleware(['auth', 'is_admin']);
 
-Route::get('/movies/{id}',"MovieController@detail");
-Route::get('/movies/{id}/edit',"MovieController@edit");
-Route::patch('/movies/{id}', 'MovieController@update');
+Route::get('/movies/{id}',"MovieController@detail")->middleware(['auth']);
+Route::get('/movies/{id}/edit',"MovieController@edit")->middleware(['auth', 'is_admin']);
+Route::patch('/movies/{id}', 'MovieController@update')->middleware(['auth', 'is_admin']);
 
+Route::get('/my-list', 'MyListController@index')->middleware('auth');
+Route::post('/my-list/{id}', 'MyListController@add')->middleware('auth');
 
+Route::get('/series', 'SerieController@index');
+Route::get('/series/new', 'SerieController@new');
+Route::post('/series/new', 'SerieController@create');
+
+Route::get('/genres', 'GenreController@index');
+Route::get('/genres/new', 'GenreController@new');
+Route::post('/genres/new', 'SerieController@create');
 
 Auth::routes();
 /*
@@ -26,7 +35,7 @@ GET  | register               | Auth\RegisterController@showRegistrationForm
 POST | register               | Auth\RegisterController@register
 */
 
-Route::get('/home', 'HomeController@index')->name('home');
+//Route::get('/home', 'HomeController@index')->name('home');
